requirements.txt
numpy
torch
pyyaml
pandas
tqdm
matplotlib
seaborn